Back-End V3/requirements.txt
aiomysql==0.3.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.3
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
databases==0.9.0
ecdsa==0.19.1
fastapi==0.121.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
